requirements.txt
# Python dependencies for DALConnector USB MIDI version
mido>=1.2.10
python-rtmidi>=1.4.9
